chore: migrate to sveltekit v2 for share
