+++ packages/share/package.json
@@ -14,8 +14,8 @@
 		"build": "svelte-kit sync && svelte-package",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-		"lint": "prettier --plugin-search-dir . --check . && eslint .",
-		"format": "prettier --plugin-search-dir . --write ."
+		"lint": "prettier --check . && eslint .",
+		"format": "prettier --write ."
 	},
 	"repository": {
 		"type": "git",
@@ -34,15 +34,17 @@
 	},
 	"homepage": "https://github.com/watergis/svelte-maplibre-components/tree/main/packages/share#readme",
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^2.1.1",
-		"@sveltejs/kit": "^1.30.4",
+		"@sveltejs/adapter-auto": "^3.1.1",
+		"@sveltejs/kit": "^2.5.4",
 		"@sveltejs/package": "^2.3.0",
-		"@typescript-eslint/eslint-plugin": "^6.21.0",
-		"@typescript-eslint/parser": "^6.21.0",
+		"@sveltejs/vite-plugin-svelte": "^3.0.2",
+		"@typescript-eslint/eslint-plugin": "^7.2.0",
+		"@typescript-eslint/parser": "^7.2.0",
 		"bulma": "^0.9.4",
 		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.35.1",
+		"pmtiles": "^3.0.5",
 		"prettier": "^3.2.5",
 		"prettier-plugin-svelte": "^3.2.2",
 		"sass": "^1.72.0",
@@ -51,14 +53,14 @@
 		"svelte-preprocess": "^5.1.3",
 		"tslib": "^2.6.2",
 		"typescript": "^5.4.2",
-		"vite": "^4.5.2"
+		"vite": "^5.1.6"
 	},
 	"type": "module",
 	"dependencies": {
 		"@fortawesome/free-solid-svg-icons": "^6.5.1",
-		"maplibre-gl": "^3.6.2",
+		"maplibre-gl": "^4.1.0",
 		"svelte-copy": "^1.4.1",
-		"svelte-fa": "^3.0.4"
+		"svelte-fa": "^4.0.2"
 	},
 	"peerDependencies": {
 		"svelte": "^4.0.0"
